Refactor qws service.
